+++ src/components/Wrappers/index.js
@@ -3,3 +3,4 @@ export { default as withServicePoints } from './withServicePoints';
 export { default as withProxy } from './withProxy';
 export { default as withDeclareLost } from './withDeclareLost';
 export { default as withClaimReturned } from './withClaimReturned';
+export { default as withMarkAsMissing } from './withMarkAsMissing';